# frontend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ILMA API
  version: 1.0.0
  description: |
    API for ILMA Education Platform (Offline-First).
    Targets CM2 students in Benin preparing for CEP exam.
    Three roles: Student, Parent, Admin.
servers:
  - url: https://api.ilma.app/v1
    description: Production
  - url: https://staging-api.ilma.app/v1
    description: Staging
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List users (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated user list
...
  /admin/users/{userId}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: User suspended }
...